test: centralize OIDC validator mock to fix security test failures
